backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
langchain>=0.0.350
langchain-community>=0.0.14
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0
unstructured>=0.10.0
beautifulsoup4>=4.12.0
requests>=2.31.0
torch>=2.0.0
transformers>=4.30.0
pydantic>=2.0.0